+++ src/console/Microsoft.Health.Fhir.Ingest.Console.Common/FhirTransformation/Processor.cs
@@ -3,10 +3,6 @@
 // Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
 // -------------------------------------------------------------------------------------------------
 
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Threading.Tasks;
 using EnsureThat;
 using Microsoft.Azure.WebJobs;
 using Microsoft.Health.Common.Telemetry;
