+++ src/console/Microsoft.Health.Fhir.Ingest.Console.Normalization/Microsoft.Health.Fhir.Ingest.Console.Normalization.csproj
@@ -14,7 +14,6 @@
 			<PrivateAssets>all</PrivateAssets>
 			<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
 		</PackageReference>
-		<PackageReference Include="Azure.Storage.Queues" Version="12.11.0" />
 	</ItemGroup>
 
 	<ItemGroup>
